revert: roll back issue #6 (form-A adapter), keep issue #5
Reverts commits 047d3a8ae and 84adb5453. The OpenAI gpt-image-2 model
does not work reliably; keeping only the Gemini/OpenRouter/AntigravityTools
chat-completions path (form B) that was verified working.

Issue #5 fixes (allowlist + tooltip + help icon alignment) are preserved.
