+++ packages/desktop/src/common/api/OpenAIRotatingClient.ts
@@ -62,15 +62,6 @@ export class OpenAIRotatingClient extends RotatingApiClient<OpenAI> {
     });
   }
 
-  async editImage(
-    params: OpenAI.Images.ImageEditParams,
-    options?: OpenAI.RequestOptions
-  ): Promise<OpenAI.Images.ImagesResponse> {
-    return await this.executeWithRetry((client) => {
-      return client.images.edit(params, options) as Promise<OpenAI.Images.ImagesResponse>;
-    });
-  }
-
   async createEmbedding(
     params: OpenAI.Embeddings.EmbeddingCreateParams,
     options?: OpenAI.RequestOptions
