+++ packages/desktop/src/common/chat/imageGenCore.ts
@@ -1,0 +1,331 @@
+/**
+ * @license
+ * Copyright 2025 AionUi (aionui.com)
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * Shared image generation logic used by both:
+ * - The built-in MCP server (imageGenServer.ts)
+ * - The legacy Gemini-specific tool (img-gen.ts)
+ */
+
+import * as fs from 'fs';
+import * as path from 'path';
+import { jsonrepair } from 'jsonrepair';
+import type OpenAI from 'openai';
+import { ClientFactory, type RotatingClient } from '@/common/api/ClientFactory';
+import type { TProviderWithModel } from '@/common/config/storage';
+import type { UnifiedChatCompletionResponse } from '@/common/api/RotatingApiClient';
+import { IMAGE_EXTENSIONS, MIME_TYPE_MAP, MIME_TO_EXT_MAP, DEFAULT_IMAGE_EXTENSION } from '@/common/config/constants';
+
+const API_TIMEOUT_MS = 120000; // 2 minutes for image generation API calls
+
+type ImageExtension = (typeof IMAGE_EXTENSIONS)[number];
+
+// ===== Utility Functions =====
+
+export function safeJsonParse<T = unknown>(jsonString: string, fallbackValue: T): T {
+  if (!jsonString || typeof jsonString !== 'string') {
+    return fallbackValue;
+  }
+
+  try {
+    return JSON.parse(jsonString) as T;
+  } catch (_error) {
+    try {
+      const repairedJson = jsonrepair(jsonString);
+      return JSON.parse(repairedJson) as T;
+    } catch (_repairError) {
+      console.warn('[ImageGen] JSON parse failed:', jsonString.substring(0, 50));
+      return fallbackValue;
+    }
+  }
+}
+
+export function isImageFile(file_path: string): boolean {
+  const ext = path.extname(file_path).toLowerCase();
+  return IMAGE_EXTENSIONS.includes(ext as ImageExtension);
+}
+
+export function isHttpUrl(str: string): boolean {
+  return str.startsWith('http://') || str.startsWith('https://');
+}
+
+export async function fileToBase64(file_path: string): Promise<string> {
+  try {
+    const fileBuffer = await fs.promises.readFile(file_path);
+    return fileBuffer.toString('base64');
+  } catch (error) {
+    const errorMessage = error instanceof Error ? error.message : String(error);
+    if (errorMessage.includes('ENOENT') || errorMessage.includes('no such file')) {
+      throw new Error(`Image file not found: ${file_path}`, { cause: error });
+    }
+    throw new Error(`Failed to read image file: ${errorMessage}`, { cause: error });
+  }
+}
+
+export function getImageMimeType(file_path: string): string {
+  const ext = path.extname(file_path).toLowerCase();
+  return MIME_TYPE_MAP[ext] || MIME_TYPE_MAP[DEFAULT_IMAGE_EXTENSION];
+}
+
+export function getFileExtensionFromDataUrl(dataUrl: string): string {
+  const mimeTypeMatch = dataUrl.match(/^data:image\/([^;]+);base64,/);
+  if (mimeTypeMatch && mimeTypeMatch[1]) {
+    const mimeType = mimeTypeMatch[1].toLowerCase();
+    return MIME_TO_EXT_MAP[mimeType] || DEFAULT_IMAGE_EXTENSION;
+  }
+  return DEFAULT_IMAGE_EXTENSION;
+}
+
+export async function saveGeneratedImage(base64Data: string, workspaceDir: string): Promise<string> {
+  const timestamp = Date.now();
+  const fileExtension = getFileExtensionFromDataUrl(base64Data);
+  const file_name = `img-${timestamp}${fileExtension}`;
+  const file_path = path.join(workspaceDir, file_name);
+
+  const base64WithoutPrefix = base64Data.replace(/^data:image\/[^;]+;base64,/, '');
+  const imageBuffer = Buffer.from(base64WithoutPrefix, 'base64');
+
+  try {
+    await fs.promises.writeFile(file_path, imageBuffer);
+    return file_path;
+  } catch (error) {
+    console.error('[ImageGen] Failed to save image file:', error);
+    throw new Error(`Failed to save image: ${error instanceof Error ? error.message : String(error)}`, {
+      cause: error,
+    });
+  }
+}
+
+// ===== Image Content Processing =====
+
+interface ImageContent {
+  type: 'image_url';
+  image_url: {
+    url: string;
+    detail: 'auto' | 'low' | 'high';
+  };
+}
+
+export async function processImageUri(imageUri: string, workspaceDir: string): Promise<ImageContent | null> {
+  if (isHttpUrl(imageUri)) {
+    return {
+      type: 'image_url',
+      image_url: { url: imageUri, detail: 'auto' },
+    };
+  }
+
+  let processedUri = imageUri;
+  if (imageUri.startsWith('@')) {
+    processedUri = imageUri.substring(1);
+  }
+
+  let fullPath = processedUri;
+  if (!path.isAbsolute(processedUri)) {
+    fullPath = path.join(workspaceDir, processedUri);
+  }
+
+  try {
+    await fs.promises.access(fullPath, fs.constants.F_OK);
+
+    if (!isImageFile(fullPath)) {
+      throw new Error(`File is not a supported image type: ${fullPath}`);
+    }
+
+    const base64Data = await fileToBase64(fullPath);
+    const mimeType = getImageMimeType(fullPath);
+    return {
+      type: 'image_url',
+      image_url: { url: `data:${mimeType};base64,${base64Data}`, detail: 'auto' },
+    };
+  } catch (error) {
+    const possiblePaths = [imageUri, path.join(workspaceDir, imageUri)].filter((p, i, arr) => arr.indexOf(p) === i);
+    const errorMessage = error instanceof Error ? error.message : String(error);
+
+    if (errorMessage.includes('Image file not found') || errorMessage.includes('not a supported image type')) {
+      throw error;
+    }
+
+    throw new Error(
+      `Image file not found. Searched paths:\n${possiblePaths.map((p) => `- ${p}`).join('\n')}\n\nPlease ensure the image file exists and has a valid image extension (.jpg, .png, .gif, .webp, etc.)`,
+      { cause: error }
+    );
+  }
+}
+
+// ===== Core Execution =====
+
+export interface ImageGenParams {
+  prompt: string;
+  image_uris?: string[] | string;
+}
+
+export interface ImageGenResult {
+  success: boolean;
+  text: string;
+  imagePath?: string;
+  relativeImagePath?: string;
+  error?: string;
+}
+
+/**
+ * Core image generation function shared between MCP server and Gemini tool.
+ */
+export async function executeImageGeneration(
+  params: ImageGenParams,
+  provider: TProviderWithModel,
+  workspaceDir: string,
+  proxy?: string,
+  signal?: AbortSignal
+): Promise<ImageGenResult> {
+  if (signal?.aborted) {
+    return { success: false, text: 'Image generation was cancelled.', error: 'cancelled' };
+  }
+
+  try {
+    // Parse image URIs
+    let imageUris: string[] = [];
+    if (params.image_uris) {
+      if (typeof params.image_uris === 'string') {
+        const parsed = safeJsonParse<string[]>(params.image_uris, null);
+        imageUris = Array.isArray(parsed) ? parsed : [params.image_uris];
+      } else if (Array.isArray(params.image_uris)) {
+        imageUris = params.image_uris;
+      }
+    }
+
+    const hasImages = imageUris.length > 0;
+    let enhancedPrompt: string;
+    if (hasImages) {
+      enhancedPrompt = `Analyze/Edit image: ${params.prompt}`;
+    } else {
+      enhancedPrompt = `Generate image: ${params.prompt}`;
+    }
+
+    const contentParts: OpenAI.Chat.Completions.ChatCompletionContentPart[] = [{ type: 'text', text: enhancedPrompt }];
+
+    // Process image URIs
+    if (hasImages) {
+      const imageResults = await Promise.allSettled(imageUris.map((uri) => processImageUri(uri, workspaceDir)));
+
+      const successful: ImageContent[] = [];
+      const errors: string[] = [];
+
+      imageResults.forEach((result, index) => {
+        if (result.status === 'fulfilled' && result.value) {
+          successful.push(result.value);
+        } else {
+          const error = result.status === 'rejected' ? result.reason : 'Unknown error';
+          const errorMessage = error instanceof Error ? error.message : String(error);
+          errors.push(`Image ${index + 1} (${imageUris[index]}): ${errorMessage}`);
+        }
+      });
+
+      successful.forEach((imageContent) => contentParts.push(imageContent));
+
+      if (successful.length === 0) {
+        return {
+          success: false,
+          text: `Error: Failed to process any images. Errors:\n${errors.join('\n')}`,
+          error: errors.join('\n'),
+        };
+      }
+    }
+
+    const messages: OpenAI.Chat.Completions.ChatCompletionMessageParam[] = [{ role: 'user', content: contentParts }];
+
+    // Create client and call API
+    const rotatingClient: RotatingClient = await ClientFactory.createRotatingClient(provider, {
+      proxy,
+      rotatingOptions: { maxRetries: 3, retryDelay: 1000 },
+    });
+
+    const completion: UnifiedChatCompletionResponse = await rotatingClient.createChatCompletion(
+      { model: provider.use_model, messages: messages as any },
+      { signal, timeout: API_TIMEOUT_MS }
+    );
+
+    const choice = completion.choices[0];
+    if (!choice) {
+      return { success: false, text: 'No response from image generation API', error: 'No response' };
+    }
+
+    const responseText = choice.message.content || 'Image generated successfully.';
+    let images = choice.message.images;
+
+    // Extract images from markdown in content if not in images field
+    if ((!images || images.length === 0) && responseText) {
+      const dataUrlRegex = /!\[[^\]]*\]\((data:image\/[^;]+;base64,[^)]+)\)/g;
+      const dataUrlMatches = [...responseText.matchAll(dataUrlRegex)];
+      if (dataUrlMatches.length > 0) {
+        images = dataUrlMatches.map((match) => ({
+          type: 'image_url' as const,
+          image_url: { url: match[1] },
+        }));
+      } else {
+        const file_pathRegex = /!\[[^\]]*\]\(([^)]+\.(?:jpg|jpeg|png|gif|webp|bmp|tiff|svg))\)/gi;
+        const file_pathMatches = [...responseText.matchAll(file_pathRegex)];
+        if (file_pathMatches.length > 0) {
+          const processedImages: Array<{ type: 'image_url'; image_url: { url: string } }> = [];
+          for (const match of file_pathMatches) {
+            const file_path = match[1];
+            const fullPath = path.isAbsolute(file_path) ? file_path : path.join(workspaceDir, file_path);
+            try {
+              await fs.promises.access(fullPath);
+              const base64Data = await fileToBase64(fullPath);
+              const mimeType = getImageMimeType(fullPath);
+              processedImages.push({
+                type: 'image_url',
+                image_url: { url: `data:${mimeType};base64,${base64Data}` },
+              });
+            } catch (_fileError) {
+              console.warn(`[ImageGen] Could not load image file: ${file_path}`);
+            }
+          }
+          if (processedImages.length > 0) {
+            images = processedImages;
+          }
+        }
+      }
+    }
+
+    if (!images || images.length === 0) {
+      const warningMessage = `Image generation did not produce any images.\n\nModel response: ${responseText}\n\nTip: Make sure your image generation model supports this type of request. Current model: ${provider.use_model}`;
+      return { success: true, text: warningMessage };
+    }
+
+    const firstImage = images[0];
+    if (firstImage.type === 'image_url' && firstImage.image_url?.url) {
+      const imagePath = await saveGeneratedImage(firstImage.image_url.url, workspaceDir);
+      const relativeImagePath = path.relative(workspaceDir, imagePath);
+
+      // Strip any inline base64 data URLs from the human-readable text before
+      // returning. The image is already saved to disk and referenced by path,
+      // so re-emitting hundreds of MB of base64 in the MCP tool response just
+      // forces the parent process to ship that payload through framed TCP again
+      // (which is where the 2026-04-14 commit-charge blow-up happened).
+      const cleanText = responseText.replace(
+        /!\[[^\]]*\]\(data:image\/[^;]+;base64,[^)]+\)/g,
+        '[embedded image extracted]'
+      );
+
+      return {
+        success: true,
+        text: `${cleanText}\n\nGenerated image saved to: ${imagePath}`,
+        imagePath,
+        relativeImagePath,
+      };
+    }
+
+    return { success: true, text: responseText };
+  } catch (error) {
+    if (signal?.aborted) {
+      return { success: false, text: 'Image generation was cancelled.', error: 'cancelled' };
+    }
+    const errorMessage = error instanceof Error ? error.message : String(error);
+    console.error(`[ImageGen] API call failed:`, error);
+    return { success: false, text: `Error generating image: ${errorMessage}`, error: errorMessage };
+  }
+}
