+++ packages/desktop/src/common/utils/imageModelAllowlist.ts
@@ -7,21 +7,19 @@
 /**
  * Allowlist for built-in image generation tool.
  *
- * Two API shapes are supported:
+ * The tool currently only supports "form B" — OpenAI chat completions multimodal
+ * output (model returns images via `message.images` or markdown). It does NOT
+ * support "form A" (`/v1/images/generations` endpoint) or async/polling APIs.
  *
- *   "chat"   — OpenAI chat completions multimodal output (form B).
- *              The model returns images via `message.images` or inline markdown.
- *              Supported: Gemini, OpenRouter chat-style, AntigravityTools.
+ * Model selection therefore must be a platform+model allowlist of providers
+ * known to work, rather than a coarse name-substring match. Otherwise users
+ * see options like `gpt-image-1` / `dall-e-3` / `sd-3.5` in the dropdown that
+ * are guaranteed to fail at runtime.
  *
- *   "images" — OpenAI /v1/images/generations + /v1/images/edits (form A).
- *              Returns { data: [{ b64_json }] }.
- *              Supported: OpenAI (gpt-image-*, dall-e-*), Stability AI,
- *              Alibaba DashScope (通义万相), Together AI (FLUX).
- *
- * When a new adapter lands, add its rule here and update the Tooltip text.
+ * Rules below mirror `useConfigModelListWithImage.ts` — the same providers we
+ * auto-supplement with default image models. When #6 lands a form-A adapter,
+ * extend this list accordingly.
  */
-
-import type { ApiShape } from '@/common/chat/imageGen/types';
 
 type ProviderShape = {
   platform?: string;
@@ -29,82 +27,27 @@ type ProviderShape = {
   name?: string;
 };
 
-const IMAGE_NAME_PATTERN = /(image|banana|imagine|flux|dall-e|stable|sd[0-9]|wanx)/i;
+const IMAGE_NAME_PATTERN = /(image|banana|imagine)/i;
 
-type Rule = {
+const RULES: Array<{
   id: string;
-  apiShape: ApiShape;
   match: (provider: ProviderShape) => boolean;
-  /** Optional additional model-name filter. If omitted, IMAGE_NAME_PATTERN is used. */
-  modelPattern?: RegExp;
-};
-
-const RULES: Rule[] = [
-  // ── Form B (chat completions) ──────────────────────────────────────────────
+}> = [
   {
     id: 'gemini',
-    apiShape: 'chat',
     match: (p) => p.platform === 'gemini' || p.platform === 'gemini-vertex-ai',
   },
   {
     id: 'openrouter',
-    apiShape: 'chat',
     match: (p) => !!p.base_url?.includes('openrouter.ai'),
   },
   {
     id: 'antigravity',
-    apiShape: 'chat',
     match: (p) => !!p.name?.toLowerCase().includes('antigravity'),
-  },
-
-  // ── Form A (OpenAI /v1/images/* endpoints) ─────────────────────────────────
-  {
-    id: 'openai-official',
-    apiShape: 'images',
-    match: (p) => !!p.base_url?.includes('api.openai.com'),
-    modelPattern: /(gpt-image|dall-e)/i,
-  },
-  {
-    id: 'stability-ai',
-    apiShape: 'images',
-    match: (p) => !!p.base_url?.includes('stability.ai'),
-    modelPattern: /(stable|sd[0-9]|ultra|core)/i,
-  },
-  {
-    id: 'dashscope',
-    apiShape: 'images',
-    match: (p) => !!p.base_url?.includes('dashscope.aliyuncs.com'),
-    modelPattern: /(wanx|image)/i,
-  },
-  {
-    id: 'together',
-    apiShape: 'images',
-    match: (p) => !!p.base_url?.includes('together') || !!p.base_url?.includes('togetherai'),
-    modelPattern: /(flux)/i,
   },
 ];
 
-/**
- * Returns true if the given provider+model combination is supported by the
- * built-in image generation tool (either form A or form B).
- */
 export const isImageGenSupported = (provider: ProviderShape, modelName: string): boolean => {
-  for (const rule of RULES) {
-    if (!rule.match(provider)) continue;
-    const pattern = rule.modelPattern ?? IMAGE_NAME_PATTERN;
-    if (pattern.test(modelName)) return true;
-  }
-  return false;
-};
-
-/**
- * Returns the API shape for a given provider: 'images' for form-A providers
- * (OpenAI /v1/images/*), 'chat' for form-B chat completions.
- * Falls back to 'chat' for unknown providers.
- */
-export const getApiShape = (provider: ProviderShape): ApiShape => {
-  for (const rule of RULES) {
-    if (rule.match(provider)) return rule.apiShape;
-  }
-  return 'chat';
+  if (!IMAGE_NAME_PATTERN.test(modelName)) return false;
+  return RULES.some((rule) => rule.match(provider));
 };
