+++ packages/desktop/src/renderer/services/i18n/i18n-keys.d.ts
@@ -1330,13 +1330,9 @@ export type I18nKey =
   | 'settings.healthStatusCleared'
   | 'settings.helpDocumentation'
   | 'settings.imageGenSupportedTooltipAntigravity'
-  | 'settings.imageGenSupportedTooltipDashscope'
   | 'settings.imageGenSupportedTooltipGemini'
-  | 'settings.imageGenSupportedTooltipOpenAI'
   | 'settings.imageGenSupportedTooltipOpenRouter'
-  | 'settings.imageGenSupportedTooltipStability'
   | 'settings.imageGenSupportedTooltipTitle'
-  | 'settings.imageGenSupportedTooltipTogether'
   | 'settings.imageGenUnsupportedTooltip'
   | 'settings.imageGeneration'
   | 'settings.imageGenerationDescription'
