+++ packages/desktop/src/renderer/services/i18n/locales/en-US/settings.json
@@ -440,14 +440,10 @@
   "mcp": "MCP Integration",
   "imageGenerationModel": "Image Model",
   "imageGenSupportedTooltipTitle": "Currently supported image models:",
-  "imageGenSupportedTooltipGemini": "Google Gemini / Vertex AI (e.g. gemini-2.5-flash-image)",
+  "imageGenSupportedTooltipGemini": "Google Gemini / Vertex AI image models (e.g. gemini-2.5-flash-image)",
   "imageGenSupportedTooltipOpenRouter": "OpenRouter chat-style image models (e.g. nano-banana, gemini-image)",
   "imageGenSupportedTooltipAntigravity": "AntigravityTools (gemini-3-pro-image)",
-  "imageGenSupportedTooltipOpenAI": "OpenAI (gpt-image-2, gpt-image-1, dall-e-3)",
-  "imageGenSupportedTooltipStability": "Stability AI (SD3.5, Stable Image Ultra/Core)",
-  "imageGenSupportedTooltipDashscope": "Alibaba DashScope — 通义万相",
-  "imageGenSupportedTooltipTogether": "Together AI — FLUX schnell / dev",
-  "imageGenUnsupportedTooltip": "Imagen and other non-OpenAI-compat APIs are not yet supported.",
+  "imageGenUnsupportedTooltip": "Models like gpt-image-* / dall-e-* / Imagen / FLUX use a different API and are not yet supported.",
   "imageGenerationGuide": "Image Model Setup Guide",
   "noAvailable": "No available image models, please configure first.",
   "mcpSettings": "MCP Tools Configuration",
