+++ packages/desktop/src/renderer/services/i18n/locales/ja-JP/settings.json
@@ -440,14 +440,10 @@
   "mcp": "MCP統合",
   "imageGenerationModel": "画像モデル",
   "imageGenSupportedTooltipTitle": "現在サポートされている画像モデル：",
-  "imageGenSupportedTooltipGemini": "Google Gemini / Vertex AI（例：gemini-2.5-flash-image）",
+  "imageGenSupportedTooltipGemini": "Google Gemini / Vertex AI 画像モデル（例：gemini-2.5-flash-image）",
   "imageGenSupportedTooltipOpenRouter": "OpenRouter 経由の chat 形式画像モデル（例：nano-banana、gemini-image）",
   "imageGenSupportedTooltipAntigravity": "AntigravityTools（gemini-3-pro-image）",
-  "imageGenSupportedTooltipOpenAI": "OpenAI（gpt-image-2、gpt-image-1、dall-e-3）",
-  "imageGenSupportedTooltipStability": "Stability AI（SD3.5、Stable Image Ultra/Core）",
-  "imageGenSupportedTooltipDashscope": "阿里 DashScope — 通義万象",
-  "imageGenSupportedTooltipTogether": "Together AI — FLUX schnell / dev",
-  "imageGenUnsupportedTooltip": "Imagen など OpenAI 非互換 API は現在サポートされていません。",
+  "imageGenUnsupportedTooltip": "gpt-image-* / dall-e-* / Imagen / FLUX などのモデルは別の API を使用するため、現在はサポートされていません。",
   "imageGenerationGuide": "画像モデル設定ガイド",
   "noAvailable": "利用可能な画像モデルがありません。先に設定してください。",
   "mcpSettings": "MCPツール設定",
