+++ packages/desktop/src/renderer/services/i18n/locales/ko-KR/settings.json
@@ -440,14 +440,10 @@
   "mcp": "MCP 통합",
   "imageGenerationModel": "이미지 모델",
   "imageGenSupportedTooltipTitle": "현재 지원되는 이미지 모델:",
-  "imageGenSupportedTooltipGemini": "Google Gemini / Vertex AI (예: gemini-2.5-flash-image)",
+  "imageGenSupportedTooltipGemini": "Google Gemini / Vertex AI 이미지 모델 (예: gemini-2.5-flash-image)",
   "imageGenSupportedTooltipOpenRouter": "OpenRouter를 통한 chat 방식 이미지 모델 (예: nano-banana, gemini-image)",
   "imageGenSupportedTooltipAntigravity": "AntigravityTools (gemini-3-pro-image)",
-  "imageGenSupportedTooltipOpenAI": "OpenAI (gpt-image-2, gpt-image-1, dall-e-3)",
-  "imageGenSupportedTooltipStability": "Stability AI (SD3.5, Stable Image Ultra/Core)",
-  "imageGenSupportedTooltipDashscope": "알리 DashScope — 통이완샹",
-  "imageGenSupportedTooltipTogether": "Together AI — FLUX schnell / dev",
-  "imageGenUnsupportedTooltip": "Imagen 등 OpenAI 비호환 API는 아직 지원되지 않습니다.",
+  "imageGenUnsupportedTooltip": "gpt-image-* / dall-e-* / Imagen / FLUX 등의 모델은 다른 API를 사용하므로 현재 지원되지 않습니다.",
   "imageGenerationGuide": "이미지 모델 설정 가이드",
   "noAvailable": "사용 가능한 이미지 모델이 없습니다. 먼저 구성하세요.",
   "mcpSettings": "MCP 도구 구성",
