+++ packages/desktop/src/renderer/services/i18n/locales/ru-RU/settings.json
@@ -432,14 +432,10 @@
   "mcp": "Интеграция MCP",
   "imageGenerationModel": "Модель для изображений",
   "imageGenSupportedTooltipTitle": "Поддерживаемые модели изображений:",
-  "imageGenSupportedTooltipGemini": "Google Gemini / Vertex AI (например, gemini-2.5-flash-image)",
+  "imageGenSupportedTooltipGemini": "Модели Google Gemini / Vertex AI (например, gemini-2.5-flash-image)",
   "imageGenSupportedTooltipOpenRouter": "Чат-модели изображений через OpenRouter (например, nano-banana, gemini-image)",
   "imageGenSupportedTooltipAntigravity": "AntigravityTools (gemini-3-pro-image)",
-  "imageGenSupportedTooltipOpenAI": "OpenAI (gpt-image-2, gpt-image-1, dall-e-3)",
-  "imageGenSupportedTooltipStability": "Stability AI (SD3.5, Stable Image Ultra/Core)",
-  "imageGenSupportedTooltipDashscope": "Alibaba DashScope — 通义万象",
-  "imageGenSupportedTooltipTogether": "Together AI — FLUX schnell / dev",
-  "imageGenUnsupportedTooltip": "Imagen и другие API, несовместимые с OpenAI, пока не поддерживаются.",
+  "imageGenUnsupportedTooltip": "Модели gpt-image-* / dall-e-* / Imagen / FLUX используют другой API и пока не поддерживаются.",
   "imageGenerationGuide": "Руководство по настройке модели изображений",
   "noAvailable": "Нет доступных моделей изображений, настройте сначала.",
   "mcpSettings": "Настройка инструментов MCP",
