+++ packages/desktop/src/renderer/services/i18n/locales/tr-TR/settings.json
@@ -440,14 +440,10 @@
   "mcp": "MCP Entegrasyonu",
   "imageGenerationModel": "Görüntü Modeli",
   "imageGenSupportedTooltipTitle": "Şu anda desteklenen görüntü modelleri:",
-  "imageGenSupportedTooltipGemini": "Google Gemini / Vertex AI (ör. gemini-2.5-flash-image)",
+  "imageGenSupportedTooltipGemini": "Google Gemini / Vertex AI görüntü modelleri (ör. gemini-2.5-flash-image)",
   "imageGenSupportedTooltipOpenRouter": "OpenRouter üzerinden chat tarzı görüntü modelleri (ör. nano-banana, gemini-image)",
   "imageGenSupportedTooltipAntigravity": "AntigravityTools (gemini-3-pro-image)",
-  "imageGenSupportedTooltipOpenAI": "OpenAI (gpt-image-2, gpt-image-1, dall-e-3)",
-  "imageGenSupportedTooltipStability": "Stability AI (SD3.5, Stable Image Ultra/Core)",
-  "imageGenSupportedTooltipDashscope": "Alibaba DashScope — 通义万象",
-  "imageGenSupportedTooltipTogether": "Together AI — FLUX schnell / dev",
-  "imageGenUnsupportedTooltip": "Imagen ve OpenAI uyumsuz diğer API'lar henüz desteklenmemektedir.",
+  "imageGenUnsupportedTooltip": "gpt-image-* / dall-e-* / Imagen / FLUX gibi modeller farklı bir API kullanır ve henüz desteklenmemektedir.",
   "imageGenerationGuide": "Görüntü Modeli Kurulum Rehberi",
   "noAvailable": "Mevcut görüntü modeli yok, lütfen önce yapılandırın.",
   "mcpSettings": "MCP Araçları Yapılandırması",
