+++ packages/desktop/src/renderer/services/i18n/locales/uk-UA/settings.json
@@ -436,14 +436,10 @@
   "mcp": "Інтеграція MCP",
   "imageGenerationModel": "Модель для зображень",
   "imageGenSupportedTooltipTitle": "Підтримувані моделі зображень:",
-  "imageGenSupportedTooltipGemini": "Google Gemini / Vertex AI (наприклад, gemini-2.5-flash-image)",
+  "imageGenSupportedTooltipGemini": "Моделі зображень Google Gemini / Vertex AI (наприклад, gemini-2.5-flash-image)",
   "imageGenSupportedTooltipOpenRouter": "Чат-моделі зображень через OpenRouter (наприклад, nano-banana, gemini-image)",
   "imageGenSupportedTooltipAntigravity": "AntigravityTools (gemini-3-pro-image)",
-  "imageGenSupportedTooltipOpenAI": "OpenAI (gpt-image-2, gpt-image-1, dall-e-3)",
-  "imageGenSupportedTooltipStability": "Stability AI (SD3.5, Stable Image Ultra/Core)",
-  "imageGenSupportedTooltipDashscope": "Alibaba DashScope — 通义万象",
-  "imageGenSupportedTooltipTogether": "Together AI — FLUX schnell / dev",
-  "imageGenUnsupportedTooltip": "Imagen та інші API, несумісні з OpenAI, поки не підтримуються.",
+  "imageGenUnsupportedTooltip": "Моделі gpt-image-* / dall-e-* / Imagen / FLUX використовують інший API і поки не підтримуються.",
   "imageGenerationGuide": "Посібник з налаштування моделі зображень",
   "noAvailable": "Немає доступних моделей зображень, спочатку налаштуйте їх.",
   "mcpSettings": "Налаштування інструментів MCP",
