+++ packages/desktop/src/renderer/services/i18n/locales/zh-CN/settings.json
@@ -440,14 +440,10 @@
   "mcp": "MCP 集成",
   "imageGenerationModel": "图像模型",
   "imageGenSupportedTooltipTitle": "当前已支持的图像模型：",
-  "imageGenSupportedTooltipGemini": "Google Gemini / Vertex AI（如 gemini-2.5-flash-image）",
+  "imageGenSupportedTooltipGemini": "Google Gemini / Vertex AI 图像模型（如 gemini-2.5-flash-image）",
   "imageGenSupportedTooltipOpenRouter": "OpenRouter 中转的图像 chat 模型（如 nano-banana、gemini-image）",
   "imageGenSupportedTooltipAntigravity": "AntigravityTools（gemini-3-pro-image）",
-  "imageGenSupportedTooltipOpenAI": "OpenAI（gpt-image-2、gpt-image-1、dall-e-3）",
-  "imageGenSupportedTooltipStability": "Stability AI（SD3.5、Stable Image Ultra/Core）",
-  "imageGenSupportedTooltipDashscope": "阿里百炼 DashScope — 通义万相",
-  "imageGenSupportedTooltipTogether": "Together AI — FLUX schnell / dev",
-  "imageGenUnsupportedTooltip": "Imagen 等非 OpenAI 兼容 API 暂不支持。",
+  "imageGenUnsupportedTooltip": "gpt-image-* / dall-e-* / Imagen / FLUX 等模型使用不同 API，暂不支持。",
   "imageGenerationGuide": "图像模型配置指南",
   "noAvailable": "暂无可用图像模型，请先配置。",
   "mcpSettings": "MCP 工具配置",
