+++ packages/desktop/src/renderer/services/i18n/locales/zh-TW/settings.json
@@ -440,14 +440,10 @@
   "mcp": "MCP 集成",
   "imageGenerationModel": "圖像模型",
   "imageGenSupportedTooltipTitle": "目前已支援的圖像模型：",
-  "imageGenSupportedTooltipGemini": "Google Gemini / Vertex AI（例如 gemini-2.5-flash-image）",
+  "imageGenSupportedTooltipGemini": "Google Gemini / Vertex AI 圖像模型（例如 gemini-2.5-flash-image）",
   "imageGenSupportedTooltipOpenRouter": "OpenRouter 轉發的圖像 chat 模型（例如 nano-banana、gemini-image）",
   "imageGenSupportedTooltipAntigravity": "AntigravityTools（gemini-3-pro-image）",
-  "imageGenSupportedTooltipOpenAI": "OpenAI（gpt-image-2、gpt-image-1、dall-e-3）",
-  "imageGenSupportedTooltipStability": "Stability AI（SD3.5、Stable Image Ultra/Core）",
-  "imageGenSupportedTooltipDashscope": "阿里百炼 DashScope — 通義萬象",
-  "imageGenSupportedTooltipTogether": "Together AI — FLUX schnell / dev",
-  "imageGenUnsupportedTooltip": "Imagen 等非 OpenAI 相容 API 暫不支援。",
+  "imageGenUnsupportedTooltip": "gpt-image-* / dall-e-* / Imagen / FLUX 等模型使用不同 API，暫不支援。",
   "imageGenerationGuide": "圖像模型配置指南",
   "noAvailable": "暫無可用圖像模型，請先配置。",
   "mcpSettings": "MCP 工具配置",
